+++ packages/timesheets/portal.manifest.mjs
@@ -1,6 +1,6 @@
 export default {
   id: 'timesheets',
-  version: '0.4.6',
+  version: '0.4.7',
   source: '@nuxt-customer-portal/timesheets',
   clientModuleId: 'timesheets',
   dependsOn: ['core', 'ui', 'organizations', 'clients'],