+++ packages/authentication/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nuxt-customer-portal/authentication
 
+## 0.4.7
+
+### Patch Changes
+
+- @nuxt-customer-portal/core@0.4.7
+- @nuxt-customer-portal/ui@0.4.7
+
 ## 0.4.6
 
 ### Patch Changes