+++ packages/invoice-products/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Product invoices
 
+## 0.4.7
+
+### Patch Changes
+
+- @nuxt-customer-portal/core@0.4.7
+- @nuxt-customer-portal/invoices@0.4.7
+- @nuxt-customer-portal/products@0.4.7
+
 ## 0.4.6
 
 ### Patch Changes